Adionada coorientação à Tcc

+++ app/Http/Controllers/SolicitacaoController.php
@@ -9,6 +9,8 @@ use App\Solicitacao;
 use App\User;
 use App\Professor;
 use App\Tcc;
+use App\Orientacao;
+use App\Coorientacao;
 
 class SolicitacaoController extends Controller
 {
@@ -27,7 +29,7 @@ class SolicitacaoController extends Controller
         $todas_solicitacoes->map(function ($solicitacao) {
             $aluno = User::where('id', $solicitacao->solicitante_id)->first();
 
-            $solicitacao['id_solicitacao'] = $solicitacao->id;
+            // $solicitacao['id_solicitacao'] = $solicitacao->id;
             $solicitacao['aluno_id'] = $solicitacao->solicitante_id;
             $solicitacao['aluno_nome'] = $aluno->name;
             $solicitacao['aluno_foto'] = $aluno->image;
@@ -44,17 +46,37 @@ class SolicitacaoController extends Controller
     {
         $solicitacao = Solicitacao::where('id', $request->solicitacao_id)->first();
         
-        // Mudar campo 'professor solicitado' no Tcc do aluno para null
-        $aluno = User::where('id', $request->aluno_id)->first();
-        $aluno->tcc->prof_solicitado = null;
-        $aluno->tcc->orientador_id = Auth::guard('professor')->user()->id;
-        
-        // Deletar solicitacao de orientacao com id informado E atualizar campos tcc do aluno
-        if($solicitacao->delete() && $aluno->tcc->save()) {
-            return redirect()->back()->with(session()->flash('info', 'Solicitação de Orientação de TCC recusada.'));
-        } 
-    
-        return redirect()->back()->with(session()->flash('error', 'Erro ao recusar Solicitação de Orientação de TCC.'));
+        if($request->tipo_solicitacao == "orientacao"){
+            // Mudar campo 'professor solicitado' no Tcc do aluno para null
+            $aluno = User::where('id', $request->aluno_id)->first();
+            $aluno->tcc->prof_solicitado = null;
+            $aluno->tcc->orientador_id = Auth::guard('professor')->user()->id;
+
+            // Criar nova orientacao
+            $orientacao = new Orientacao;
+            $orientacao->orientador_id = Auth::guard('professor')->user()->id;
+            $orientacao->aluno_id = $request->aluno_id;
+                    
+            // Deletar solicitacao de orientacao com id informado, criar novo orientacao e atualizar campos tcc do aluno
+            if($solicitacao->delete() && $aluno->tcc->save() && $orientacao->save()) {
+                return redirect()->back()->with(session()->flash('info', 'Solicitação de Orientação de TCC aceita.'));
+            } 
+
+            return redirect()->back()->with(session()->flash('error', 'Erro ao aceitar Solicitação de Orientação de TCC.'));
+            
+        } else if ($request->tipo_solicitacao == "coorientacao") {
+
+            // Criar nova coorientacao
+            $coorientacao = new Coorientacao;
+            $coorientacao->coorientador_id = Auth::guard('professor')->user()->id;
+            $coorientacao->aluno_id = $request->aluno_id;
+                    
+            if($solicitacao->delete() && $coorientacao->save()) {
+                return redirect()->back()->with(session()->flash('info', 'Solicitação de Coorientação de TCC aceita.'));
+            } 
+            
+            return redirect()->back()->with(session()->flash('error', 'Erro ao aceitar Solicitação de Coorientação de TCC.'));
+        }
 
     }
 
@@ -62,16 +84,24 @@ class SolicitacaoController extends Controller
     {
         $solicitacao = Solicitacao::where('id', $request->solicitacao_id)->first();
         
-        // Mudar campo 'professor solicitado' no Tcc do aluno para null
-        $aluno = User::where('id', $request->aluno_id)->first();
-        $aluno->tcc->prof_solicitado = null;
-        
-        // Deletar solicitacao de orientacao com id informado E atualizar campo tcc do aluno
-        if($solicitacao->delete() && $aluno->tcc->save()) {
-            return redirect()->back()->with(session()->flash('info', 'Solicitação de Orientação de TCC recusada.'));
-        } 
+        if($request->tipo_solicitacao == "orientacao"){
+            // Mudar campo 'professor solicitado' no Tcc do aluno para null
+            $aluno = User::where('id', $request->aluno_id)->first();
+            $aluno->tcc->prof_solicitado = null;
+            
+            // Deletar solicitacao de orientacao com id informado E atualizar campo tcc do aluno
+            if($solicitacao->delete() && $aluno->tcc->save()) {
+                return redirect()->back()->with(session()->flash('info', 'Solicitação de Orientação de TCC recusada.'));
+            } 
+            return redirect()->back()->with(session()->flash('error', 'Erro ao recusar Solicitação de Orientação de TCC.'));
+
+        } else if ($request->tipo_solicitacao == "coorientacao") {
+            if($solicitacao->delete()) {
+                return redirect()->back()->with(session()->flash('info', 'Solicitação de Coorientação de TCC recusada.'));
+            }
+            return redirect()->back()->with(session()->flash('error', 'Erro ao recusar Solicitação de Coorientação de TCC.'));
+        }
     
-        return redirect()->back()->with(session()->flash('error', 'Erro ao recusar Solicitação de Orientação de TCC.'));
     }
     
 }
